+++ NEWS.md
@@ -33,6 +33,8 @@
   to better comply with the LRM.
 - Fixed a crash when string literal characters have a type which is an
   alias to another type (#801).
+- Added a warning when calling the predefined `"="` and `"/="` operators
+  on arrays and the left and right hand sides have different lengths.
 
 ## Version 1.10.4 - 2023-10-16
 - Fixed compatibility with LLVM 17.